refactor:PeopleTILSection Empty 컴포넌트 루트 프로토타입에 포함
