+++ src/components/roadmap/PeopleTIL/PeopleTILSection/index.tsx
@@ -1,4 +1,8 @@
+import Image from 'next/image';
+import { useRouter } from 'next/router';
+import Button from '@/components/common/Button';
 import TIL from '@/components/roadmap/PeopleTIL/TIL';
+import { tilyLinks } from '@/constants/links';
 import * as Styled from './style';
 
 const PeopleTILSection = () => {
@@ -17,3 +21,26 @@ const PeopleTILSection = () => {
 };
 
 export default PeopleTILSection;
+
+PeopleTILSection.Empty = function Empty() {
+  const router = useRouter();
+
+  return (
+    <Styled.Root>
+      <Styled.Title>다른 사람의 TIL 보기</Styled.Title>
+      <Styled.CardContainer>
+        <Styled.ImageContainer>
+          <Image src="/assets/icons/ic_peopleTILEmpty.svg" width={200} height={200} alt="다른 사람의 TIL이 없습니다." />
+        </Styled.ImageContainer>
+
+        <Styled.Description>
+          <span>공개된 다른 TIL이 없습니다.</span>
+        </Styled.Description>
+
+        <Button css={Styled.ButtonStyles} onClick={() => router.push(tilyLinks.home())}>
+          메인 페이지로 이동
+        </Button>
+      </Styled.CardContainer>
+    </Styled.Root>
+  );
+};
