+++ src/pages/roadmap/[roadmapId]/step/[stepId].tsx
@@ -1,6 +1,5 @@
 import styled from '@emotion/styled';
 import HeaderLayout from '@/components/layout/HeaderLayout';
-import EmptyPeopleTILSection from '@/components/roadmap/PeopleTIL/EmptyPeopleTILSection';
 import FeatureInfoSection from '@/components/roadmap/PeopleTIL/FeatureInfoSection';
 import PeopleTILSection from '@/components/roadmap/PeopleTIL/PeopleTILSection';
 import { setLayout } from '@/utils/layout';
@@ -12,7 +11,7 @@ const PeopleTil = () => {
         <Inner>
           <FeatureInfoSection />
           {/* <PeopleTILSection /> */}
-          <EmptyPeopleTILSection />
+          <PeopleTILSection.Empty />
         </Inner>
       </Root>
     </>
